x/energija/client/cli/query_energy_store.go
package cli

import (
	"context"

	"github.com/spf13/cobra"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/kiprasdak/energija/x/energija/types"
)

func CmdListEnergyStore() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-energy-store",
		Short: "list all energy-store",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllEnergyStoreRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.EnergyStoreAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowEnergyStore() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-energy-store [index]",
		Short: "shows a energy-store",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetEnergyStoreRequest{
				Index: argIndex,
			}

			res, err := queryClient.EnergyStore(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
